+++ src/phishbench/input/email_input/models/_body.py
@@ -74,6 +74,7 @@ def decode_text_part(part):
 
 def clean_html(raw_html: str):
     assert isinstance(raw_html, str)
+    # pylint: disable=no-name-in-module
     cleaner = Cleaner()
     # We only want to remove javascript and style
     cleaner.javascript = True
@@ -102,6 +103,7 @@ def clean_html(raw_html: str):
 class EmailBody:
     """
     A class representing the body of an email.
+
     Attributes
     ----------
     text : str