+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change Log
 
+## Major changes
+
+* `phishbenh.input.email_input.read_email_from_file` now returns an `EmailMessage` instead of a raw `Message` object.
+
 # v2.0.1
 
 ## Major changes